requirements.txt
Flask
Flask-Cors
pycryptodome
